requirements.txt
django
qrcode
python-barcode
pillow
gunicorn
whitenoise
